Apply the same fix from the campaign subscriber to the token subscriber
