+++ EventListener/CampaignSubscriber.php
@@ -32,14 +32,13 @@ use MauticPlugin\CustomObjectsBundle\Helper\QueryFilterHelper;
 use MauticPlugin\CustomObjectsBundle\Repository\DbalQueryTrait;
 use Doctrine\DBAL\Connection;
 use MauticPlugin\CustomObjectsBundle\Entity\CustomItem;
-use MauticPlugin\CustomObjectsBundle\Repository\DbalQueryBuilderParamCopyTrait;
 use MauticPlugin\CustomObjectsBundle\Entity\CustomField;
-use Mautic\LeadBundle\Segment\Query\QueryBuilder;
+use MauticPlugin\CustomObjectsBundle\Helper\QueryBuilderManipulatorTrait;
 
 class CampaignSubscriber extends CommonSubscriber
 {
     use DbalQueryTrait;
-    use DbalQueryBuilderParamCopyTrait;
+    use QueryBuilderManipulatorTrait;
 
     /**
      * @var TranslatorInterface
@@ -252,37 +251,6 @@ class CampaignSubscriber extends CommonSubscriber
             $event->setResult(true);
         } else {
             $event->setResult(false);
-        }
-    }
-
-    /**
-     * Empty and NotEmpty operators require different/opposite behaviour than what segment helper does.
-     * We have to handle it ourselves here.
-     *
-     * @param string       $operator
-     * @param string       $queryAlias
-     * @param QueryBuilder $innerQueryBuilder
-     */
-    private function handleEmptyOperators(string $operator, string $queryAlias, QueryBuilder $innerQueryBuilder): void
-    {
-        if ('empty' === $operator) {
-            $innerQueryBuilder->resetQueryPart('where');
-            $innerQueryBuilder->where(
-                $innerQueryBuilder->expr()->orX(
-                    $innerQueryBuilder->expr()->isNull($queryAlias.'_value.value'),
-                    $innerQueryBuilder->expr()->eq($queryAlias.'_value.value', $innerQueryBuilder->expr()->literal(''))
-                )
-            );
-        }
-
-        if ('!empty' === $operator) {
-            $innerQueryBuilder->resetQueryPart('where');
-            $innerQueryBuilder->where(
-                $innerQueryBuilder->expr()->andX(
-                    $innerQueryBuilder->expr()->isNotNull($queryAlias.'_value.value'),
-                    $innerQueryBuilder->expr()->neq($queryAlias.'_value.value', $innerQueryBuilder->expr()->literal(''))
-                )
-            );
         }
     }
